Admin/Errors: Adds backtrace to DB. Ability to view details on certain error

+++ admin/errors.php
@@ -3,8 +3,10 @@
 /**
  * erreurs.php
  *
- * @version 1.1  - Remade with more robust template by Gorlum for http://supernova.ws
- * @version 1.0s - Security checked for SQL-injection by Gorlum for http://supernova.ws
+ * @version 1.2   - Added ability to view single error details (like backtrace) by Gorlum for http://supernova.ws
+ * @version 1.1st - Tested by Gorlum for http://supernova.ws
+ * @version 1.1   - Remade with more robust template by Gorlum for http://supernova.ws
+ * @version 1.0s  - Security checked for SQL-injection by Gorlum for http://supernova.ws
  * @version 1.0
  * @copyright 2008 by e-Zobar for XNova
  */
@@ -21,6 +23,7 @@ includeLang('admin');
 $parse = $lang;
 
   $delete    = intval($_GET['delete']);
+  $detail    = intval($_GET['detail']);
   $deleteall = SYS_mysqlSmartEscape($_GET['deleteall']);
 
   if ($user['authlevel'] >= 3) {
@@ -32,24 +35,34 @@ $parse = $lang;
       doquery("TRUNCATE TABLE `{{table}}`", 'errors');
     }
 
-    // Afficher les erreurs
-    $query = doquery("SELECT * FROM `{{table}}`", 'errors');
-    $i = 0;
-    while ($u = mysql_fetch_array($query)) {
-      $i++;
-      $parse['errors_list'] .= "
-      <tr><td class=n>". $u['error_id'] ."</td>
-      <td class=n>". $u['error_sender'] ."</td>
-      <td class=n>". $u['error_type'] ."</td>
-      <td class=n>". date($config->game_date_withTime, $u['error_time']) ."</td>
-      <td class=b>". $u['error_page'] ."</td>
-      <td class=n><a href=\"?delete=". $u['error_id'] ."\"><img src=\"../images/r1.png\"></a></td>
-      </tr>
-      <tr><td colspan=\"6\" class=b>".  nl2br($u['error_text'])."</td></tr>";
-    }
-    $parse['errors_num'] = $i;
+    if($detail){
+      $errorInfo = doquery("SELECT * FROM `{{table}}` WHERE `error_id` = {$detail}", 'errors', true);
 
-    display(parsetemplate(gettemplate('admin/errors_body'), $parse), "Bledy", false, '', true);
+      $parse = array_merge($parse, $errorInfo);
+      $parse['error_time'] = date($config->game_date_withTime, $errorInfo['error_time']);
+
+      display(parsetemplate(gettemplate('admin/error_detail'), $parse), "Errors", false, '', true);
+    }else{
+      // Afficher les erreurs
+      $query = doquery("SELECT * FROM `{{table}}`", 'errors');
+      $i = 0;
+      while ($u = mysql_fetch_array($query)) {
+        $i++;
+        $parse['errors_list'] .= "
+        <tr><td class=n><a href=errors.php?detail={$u['error_id']}><u>{$u['error_id']}</u></a></td>
+        <td class=n>{$u['error_sender']}</td>
+        <td class=n>{$u['error_type']}</td>
+        <td class=n>". date($config->game_date_withTime, $u['error_time']) ."</td>
+        <td class=b>{$u['error_page']}</td>
+        <td class=n><a href=\"?delete=". $u['error_id'] ."\"><img src=\"../images/r1.png\"></a></td>
+        </tr>
+        <tr><td colspan=\"6\" class=b>".  nl2br($u['error_text'])."</td></tr>";
+      }
+      $parse['errors_num'] = $i;
+
+      display(parsetemplate(gettemplate('admin/errors_body'), $parse), "Errors", false, '', true);
+    }
+
   } else {
     message( $lang['sys_noalloaw'], $lang['sys_noaccess'] );
   }
